Fix bug 1397377: Restore ability to delete translations (#1152)
Translation can only be deleted, if it's rejected. That means we're
implicitly introducing a confirmation step for any non-rejected
translation. If you want to delete an approved or unreviewed
translation, you need to reject it first, and then delete it.

This approach also keeps Approve & Reject as the only first-rate
actions. The Delete action should be used in edge cases, more rarely,
and is hence, discouraged. 

Other changes:
* Make <menu> height always the same, even for non-authenticated users
  and read-only translations. That makes the helpers height occupy the
  entire available height across all pages.
* Unify the way of getting translations from the DB.
* Improve some comments.