Fix bug 1490351 - Add actions on status icons in History tab. (#1137)
This enables approve, reject, unapprove and unreject actions, accessible from the status icons next to each translation in the History tab. It correctly updates the status on the server and the local state. It also creates a new view dedicated to approving translations, in order to not have to use update_translation, which requires a lot of parameters for just changing a status. It refreshes History data from the server, to make sure that the data is correct instead of trying to fix it locally.