Fix bug 1497572: Improve custom deployment experience  (#1181)
After initial Pontoon set up, the /teams and /projects page 404. With
this patch we display a quick notice that no projects have been set
up yet with a link to set them up.

Relatedly, the "Start Localizing Now" CTA link on the homepage is broken
and this patch fixes it.