Fix bug 1490346 - Copy translations into the Editor on click.
This enables copying a translation in one of the helper tabs (History, Machinery, Locales) into the Editor when clicked. It involves refactoring some of those components to make the data flow the right way. It might not be the best way to do this, but it works and is satisfying for now. We might want to refactor this part of the code again later.

This also separates CSS rules for the various Translation components of the tabs, so that they do not impact each other.