Fix bug 1527911 - Allow users to switch between Translate apps (#1203)
* Support `all-resources` value for the resource parameter.

This makes Translate.Next accept the same URL patterns as the current app, making it easier to transition from one to the other.

* Add a user preference to decide which Translate app to load.

This preference will be used to redirect a user to the current Translate page or the new Translate.Next app.

* Add redirections based on the user setting use_translate_next.

If the user has use_translate_next set to True, then when loading the old translate page, they will get redirected to Translate.Next instead. The opposite goes if the setting is set to False.

* Add links to switch between the two Translate apps.

Clicking those link on each page will update the user preference and then redirect to the other page.