Bug 1492166 - Add a tool to mark patterns in a string.
This tool can be used to generate a React component that will apply a list of parsers to a string, and replace each matching pattern or term with a special marker. It currently supports RegExp and simple string terms. The markers can be anything, allowing us to define specific tags to use for each parser.

Note that this tool is in a `lib` folder because I intend to extract it to a library in the future.