Fixes bug 1492166 - Add placeables to Metadata and helper tools.  (#1233)
This marks placeables in the Metadata, and makes each one clickable. On click, the matched content of the placeable is added to the Editor. A big refactoring was needed to be able to do that. When text is selected in the Editor, that text is replaced with the placeable content.

This also marks placeables in all three helper tabs (History, Machinery and Locales). Because some of those use diffs, this added a way to combine our WithPlaceables and TranslationDiff. It also creates a new list of parsers to use specifically with diff, to remove false positives that would come from the leading space parser.