Fix bug 1519894: Add Resource menu
Exposes the name of the resource currently localized in the main
navigation, and lists the entire batch of project resources with their
completion percentages in the corresponding menu.

Selecting a resource in the menu hot-reloads the translate view. Before
loading we check for any unsaved changes in the editor, which we now
also do upon clicking on the resource path in the Metadata section.