No bug - Split entitieslist module. (#1305)
This moves the data part (reducer, selectors, actions) of the entitieslist module into a core/entities module, and moves the types into the api module. The goal is to make imports around the app more consistent (we should not import from the modules/ folder, only from core/). I also want to consolidate types for data that comes from the API into the api module, and this is one step towards that goal. This is purely a refactoring, no functionalities have changed.