Fix bug 1556771 - Show a simple text editor for simple Fluent messages. (#1325)
This commit refactors the whole `editor` module into 3 smaller modules: 
- `modules/genericeditor` has code for the generic editor, the one we show for most formats;
- `modules/fluenteditor` has code specific for the Fluent format, including a source editor and all variants of the rich Fluent editor;
- `core/editor` has code that is shared by both modules, including a `connectedEditor` higher-order component that has all the data and methods Editor implementations need.

It also adds basic support for the Fluent Rich Editor. It allows having multiple translation forms for a single string, and switching between the "rich" mode and the "source" mode. It adds a "simple" rich editor for all strings that are, well, "simple".