Fix bug 1510185: Add extra filters (#1343)
In addition to fixing the bug, this patch also:
* Fixes the size of the Unreviewed checkbox
* Shows checkmark when hovering filter icons
* Brings some refactorings to better support multiple groups of filters

It also drops the Untranslated filter due to low usage compared to the
code and UX complexity it comes with. The logs only show 10 hits per day,
and it's impossible to tell how many of those are triggered by the
Untranslated filter and how many are manual selections of the
"Fuzzy, Errors, Missing" filters.