Fix bug 1566869: Reset ignored state on unsavedchanges update (#1358)
When a translation is saved or added, we set the "ignored" boolean to true in unsavedchanges to hide the check before moving to the next entity. However, that boolean was never reset, because that only happened when "hide" was called, and hide could only be called when the panel was displayed, and the panel was never displayed because ignored was true. This patch resets the "ignored" state to false every time the state is updated.