Fix bug 1575875 - Add support for TextElements in Fluent Rich Editor (#1355)
This adds a new Rich Editor component that displays Fluent messages in a nicer, user-friendly UI. It currently only supports Fluent messages that only contain simple text elements or patterns. 

The main change is that the translation can now be stored as either a string or a Fluent AST object. That AST form is only used by the Rich Editor, but this requires adding a bunch of checks all around the place, for the sake of Flow notably.