Fix bug 1583776 - Remove Translate.Next switch. (#1403)
This patch removes the user setting controlling whether Translate.Next is enabled or not, and instead use only the waffle flag to choose if Translate.Next or translate current should be shown to the user. It also replaces the link to switch in the front-end with a message and link to our release announcement blog post.