Fix bug 1504702 - Add pseudo localization tool for development. (#1407)
* Fix bug 1504702 - Add pseudo localization tool for development.

This adds support for two kinds of pseudolocalization, "accented" and "bidi", to the Translate app. They can be enabled by adding `pseudolocalization=[accented|bidi]` to the URL. When enabled, only the en-US strings will be loaded, and they will be transformed before being rendered. 

This allows to spot what strings are not currently localized in code, and can show problems with string size, support of unicode characters, etc.