Fix bug 1552801: [pretranslation] Implement core task (#1310)
The task identifies strings without any translations and any suggestions,
and uses a simple algorithm to gather pretranslations, which are saved as
suggestions.

The algorithm returns the first 100% match it finds in translation memory.
If it doesn't find any, it engages Google Translate. Translation user is
set to newsly created TM and GT users correspondingly.

The task cannot be triggered yet, bug 1552802 will take care of that. We have
already added a pretranslation_enabled toggle to the Project model, though.

The majority of code is hosted in the newly created pontoon.pretranslation app.