No bug - Allow changing files from volumes in docker. (#1502)
When creating containers with docker-compose, volumes containing the source code from the host are created. From inside a docker container, it was not possible to make changes in those volumes, because of a permission issue. This commit sets the GID of the pontoon group so that it matches what will, in most cases, be the group of the user in the host (1000). That allows to make changes inside volumes. Note that if a user has a different group GID (for example, 1001, because they are the second user created on the host) then this will not work. But most folks work on their personal computer so that should not happen too often, and I do not have a good solution for that anyway. This is strictly better than before.