Bug 1599075 - Show Approve button when translation already exists. (#1507)
* Bug 1599075 - Show Approve button when translation already exists.
This makes the actions behind the "Save" button of the Editor clearer: only two UI states where shown so far, but in reality three kinds of actions could happen. The "Approve" action was hidden, and this commit makes it explicit by changing the button label to show what is going to happen in the back-end. This also makes the front-end control the behavior instead of relying on the back-end, allowing us to make more focus back-end views in a follow-up.

This also implements the same behavior for the "Enter" shortcut of the Editor, and fixes an issue with FailedChecks where "Save anyway" was being shown instead of "Suggest anyway" in some cases.