Fix bug 1602973 - Make sure users cannot send the same translation twice. (#1521)
This prevents users from spamming the "save" button, which if done fast enough can lead to the same translation being sent several times. Now as soon as the user clicks the Save button, it becomes disabled until the `sendTranslation` action is entirely finished. I am also using this opportunity to introduce spinners and label update on buttons, as a proposal. It is a pattern that I like, so I would like to generalize it on the app.