No bug - Add a guide to first Pontoon contribution. (#1532)
This adds a new guide to start contributing to Pontoon, detailing all the needed steps and giving pointers to useful documentation. It also reworks the documentation home page to be more helpful to people depending on what they are looking for. And it does the same for the readme file.