No bug - Add false migrations triggered by black. (#1559)
While running black on our entire code base, our models got affected. That was only cosmetic, of course, but django is not able to detect that, and think we made actual changes to them. My guess is that it's mostly because of text content that has been reformatted. The simplest way to deal with that is simply to run those pesky migrations, and ignore them later on.