Bug 1527853 - Refactor update_translation view (#1543)
This commit removes the `update_translation` view from the base app, and moves it into a new `translations` app as a more focused `create_translation` view. It also merges the `review` app into the new `translations` app, as it made no sense to have those separated (as they are both related to the Translation model).