Fix bug 1611666 - Prevent loss of letters while typing in the search input. (#1569)
There was a concurrency issue created by the fact that the search input content was being loaded from the URL params every time those changed. That meant that, if the user was typing at just the right time, they would 1. update the search state variable 2. trigger a search after a delay, which updates the URL 3. update the search state variable with a new letter 4. the search content gets set to the previous state because of the URL update, erasing the last letter the user typed. This patch fixes this problem by only updating the search state variable on mount or if it is empty in the new URL.