Fix bug 1616918 - Correctly compare with Fluent strings in history (#1574)
This solves a few comparison problems when looking for a Fluent translation in the history. It makes sure to turn all FTL string in Fluent messages, to use the `.equals` method and make sure comparison works correctly. This avoid bugs when a Fluent string misses the trailing newline character, for example. It also fly-by solves issues related to the unsaved changes checks.