Fix bug 1629586: Switch order of Select All & N Strings Selected
After selecting an entity for batch editing, the batch editor opens. If you want
to select all strings for batch editing, you need to move the mouse all the way
to the opposite site of the screen.

By switching the Select All button and N Strings Selected title, that path becomes
much shorter.