Improve performance when typing in Editor.
This does two things. First, it removes the dependency on the editor state from EntityDetails, removing the need to re-render that (big) part of the page every time a user types. To do that, this transforms the machinery Translation component into a functional component using hooks.

Second, it changes the way components get `unsavedchanges` data, to re-render only when either `exist` or `ignored` change. This once again allows to not re-render a huge chunk of the page (most notably, the entities list) on each change of the editor content.