Fix bug 1633149: Detect changes to translated resources on sync
When detecting changes to translated resources in VCS, we need to
identify which of the changed paths belongs to a localizable resource.

The logic that does that makes assumptions about the file and folder
structure which are not valid for projects using project configuration.

Hence this patch adds a new logic for such projects, which makes use of
the ProjectFiles matcher of the compare-locales library.

The logic also converts the changed paths to the canonical form for
representing resources used across Pontoon codebase (reference path
insted of l10n path) and assigns the list of changed locales to each
path.

This patch also fixes bug 1608045, which was another symptom of the
very same problem.