Fix bug 1688573: reduce renundant re-renders caused by typing (#1854)
Because typing constantly updates `editor.translation`, the idea is to
subscribe to its changes as less as possible. When `editor.translation`
must inevitably be read but the data won't be used as-is but rather
derived into some other piece of data, the calculation is done behind a
selector to avoid extra re-renders.

Likewise, the `withDiff` HoC has been wrapped around `React.memo`,
because rendering a diff can take ~3ms, and with a higher number of
History/Helpers entries, rendering of every entry will otherwise be
considerably slower.