Rename module to ensure import resolution works
tsc would resolve 'history' to the same module where the import happens
instead of resolving to the 'node_modules/history/' module.

A possible way to overcome that is to set `basePath: .` and then set
individual `paths: { ... }` mappings to support absolute imports within
the project. Unfortunately that's not supported by CRA and it'll even
overwrite such modifications when building. So the easiest solution is
to rename this module and avoid the problem altogether.