Prevent app crash on Heroku by forcing protobuf to use Python for parsing
By default, protobuf uses C++ for parsing. The library includes prebuilt binary modules for Apple silicon, which my local box uses, so for me the
library works as expected without this hack. I only noticed that's not the case for everyone after deploying to the stage server on Heroku.

Read more:
https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates
https://developers.google.com/protocol-buffers/docs/reference/python-generated#sharing-messages