Include all source string translations in localized XLIFF files (#2871)
The problem surfaced when a new string was added to the reference file, but not the the localized file. When it got translated, it was missing in the exported localized XLIFF file.

XLIFF files are now handled more like asymmetric files on serialization, which means we use the source file as a template and add target note with translations when available.