Refactor data gathering for the /insights page
1. Fix #2979: Optimize /insights page.

To render the charts on the Insights pages, we use periodically gathered ProjectLocaleInsightsSnapshot data. That works fine on Project, Locale and
ProjectLocale pages, but not on the /insights page which shows data across the entire app. The amount of data requested from the DB quickly becomes big and the request starts timing out.

In this patch, we gather data directly from the ActionLog as hinted at in #2938.

2. Fix #2999: Show insights for all months.

3. Calculate monthly averages rather than averages of averages.