Optimize filters by introducing Resource.order
String filters can be slow, especially because ordering entities by Resource.path is slow.

It turns out that:
- Adding DB index to Resource.path doesn't bring performance gains.
- Sorting entities by any other Resource.field is fast.

The reason for slowness is probably that Resource.path can be a pretty long text field.

So we're now introducing Resource.order, which to represent the alphabetic order of resources in the project.

We reorder resources when new resources are added to the project.