Use proper paging for entities (#3129)
We currently use a weird way of paginating entities - in each request we include a list of already loaded entities, which we exclude from the matching entities in Django and then return the first page of the remaining entities. This is a relic of the react-infinite-scroller library which we no longer use.

We should implement proper paging, i.e. drop exclude_entities and only return the relevant page.

That should also speed up loading of entities, because NOT IN DB queries tend to be slow and because less data will be transferred.