Do not exclude specific TM entries
Each request to fetch Translation Memory matches excludes TM entries that belong to Translations that are Fuzzy or not-Approved.
That is not necessary, because we remove TM entries when translations get Unapproved, Rejected or become Fuzzy.
It could also speed up Translation Memory requests, because join with the Translation table is no longer needed.

Also included is a data migration to remove any legacy TM entries
that belong to Translations that are Fuzzy or not-Approved.