Keep cache warm (#3168)
We cache data for some of the views (e.g. Contributors) for a day. Some of them don't get a lot of visits, not even one per day, meaning that the visitors of these pages often hit the cold cache.

This command refreshes data in the cache every day, because it changes often. The command is designed to run daily.