Use default Gunicorn worker timeout setting
The Heroku router drops a long-running request after 30 seconds, but the dyno behind it
continues processing the request until completion.

Adding a timeout to the webserver ensures that the dyno itself drops the long-running request,
creating capacity for other requests.

A while ago the webserver timeout was set to 120 seconds, in order to potentially help
with long request like downloading TMX files via curl. It's not clear if this was actually
helpful, but it increases a chance of Applcation Errors under heavy load.