Add banner for project managers, given priority to role within locale
This adds a banner for users defined as "Project Manager" within a project.
To reduce the space for confusion, I changed the tooltip of MNGR from "Manager" to "Team Manager".

This also makes the priority of roles consistent between front-end and back-end:
- If a user as a role within the locale (translator, manager), we use that for the banner.
- If a user is set as PM, we use that even if the user is an Admin.

Finally, this adds CSS variables for users, instead of reusing the ones for translation status.