Show badges on profile page and add logic to award them on relevant activities (#3423)
Display badges in the user profile page. The respective badges are displayed if the user's level for that badge is above 1, and the level of the badge is displayed on the bottom left of the badge.

Adds the relevant stat tracking needed to award badges based on the criteria defined in the specification.