Only show local part of the email address in the Profile page title
The motivation for this change is that we currently publicly show the email address of a user that doesn't have any contributions yet, but our Terms of Use say "Your name and email address, or a derivative of it, may be attached to your *contributions* and so be visible worldwide".

Note that we show the custom display name in the Profile page title for users that set it.