Add ability to send Monthly activity summary (#3474)
The changeset also factors out the base HTML email template, which is currently shared with the notifications email template. In the future, we'll use it for all HTML emails we send.

In addition to the spec, a configurable Thank you note is added. The day of the month to send the summary emails on is also configurable.