Feature #3605: Displays Notification "real" date instead of nnn days/weeks ago
Reading the "19 days/weeks ago" on a Notification component in
the Notifications page is difficult to visualize. This addition
makes it such that once a Notification becomes older than 7
days, the message displayed is a date in the form of "YYYY-MM-DD"
instead.

This change only affects the main Notifications page.