Use MF2 syntax for plural gettext messages (#3635)
This is an intermediate step towards #3538, de-specializing gettext .po plural message handling.

Previous to this PR, each gettext plural message had its source plural form stored as Entity.string_plural, and for each target locale, each of its variants was separately translated, with Translation.plural_form indicating the variant identifier. These fields are not used by any other format, and they are removed from the database in this change.

To represent plural source and target messages, Unicode MessageFormat (MF2) syntax representation is now used internally for all gettext messages, but this is never revealed to users. Placeholders like {0} or %(type)s are not parsed as such, but treated as a literal part of the string.

The translator front-end is modified to use the "rich editor" also for gettext plurals; this was previously only used by Fluent messages with variants. The FTL source view is not enabled for gettext plurals.

Also:
* Drop unused Locale.cldr_* methods + simplify & fix what remains. Previously, Locale.cldr_plurals_list() would return ['one'] if cldr_plurals was empty. This is clearly wrong, and it now returns ['other'].
* Disable pretranslation for plural gettext messages (see #3694)